Reorganizing based on app.routing.ts
