+++ src/app/app.module.ts
@@ -14,14 +14,14 @@ import { AngularMaterialModule } from './angular-material.module';
 
 import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app.routing';
-import { DpkFormComponent } from './components/dpk-form/dpk-form.component';
-import { DpkFormService } from './components/dpk-form/dpk-form.service';
-import { LoginComponent } from './components/login/login.component';
+import { DpkFormComponent } from './components/sidenav/dpk-form/dpk-form.component';
+import { DpkFormService } from './components/sidenav/dpk-form/dpk-form.service';
+import { LoginComponent } from './components/sidenav/login/login.component';
 import { PdfSlideComponent } from './components/dpk/pdf-slide/pdf-slide.component';
 import { SidenavComponent } from './components/sidenav/sidenav.component';
-import { SliderItemsComponent } from './components/slider-items/slider-items.component';
-import { SliderItemDirective } from './components/slider-items/slider/slider-item.directive';
-import { SliderComponent } from './components/slider-items/slider/slider.component';
+import { SliderItemsComponent } from './components/sidenav/slider-items/slider-items.component';
+import { SliderItemDirective } from './components/sidenav/slider-items/slider/slider-item.directive';
+import { SliderComponent } from './components/sidenav/slider-items/slider/slider.component';
 import { DpkParseService } from './components/dpk/slides/dpk-parse.service';
 import { SlideComponent } from './components/dpk/slides/slide/slide.component';
 import { SlidesComponent } from './components/dpk/slides/slides.component';
