+++ src/app/components/dpk/pdf-slide/pdf-slide.component.css
