+++ src/app/components/dpk/singer-view/singer-view.component.css
