+++ src/app/components/dpk/slides/slide/slide.component.css
