+++ src/app/components/dpk/slides/slide/slide.component.ts
@@ -1,12 +1,12 @@
 import { Component, HostListener, Input, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Observable } from 'rxjs';
-import { fadeAnimation } from '../../../animations/fade.animation';
-import { FirebaseBhajan } from '../../../interfaces/bhajan';
-import { DriveImageList } from '../../../interfaces/drive';
-import { SlideConfigI } from '../../../interfaces/slide-config-i';
-import { DriveAPIService } from '../../../services/drive-api.service';
-import { SlideService } from '../../../services/slide.service';
+import { fadeAnimation } from '../../../../animations/fade.animation';
+import { FirebaseBhajan } from '../../../../interfaces/bhajan';
+import { DriveImageList } from '../../../../interfaces/drive';
+import { SlideConfigI } from '../../../../interfaces/slide-config-i';
+import { DriveAPIService } from '../../../../services/drive-api.service';
+import { SlideService } from '../../../../services/slide.service';
 import { DpkParseService } from '../dpk-parse.service';
 
 @Component({
