+++ src/app/components/dpk/slides/slides.component.html
