+++ src/app/components/dpk/slides/slides.component.spec.ts
