+++ src/app/components/sidenav/slider-items/slider-items.component.spec.ts
