+++ src/app/components/sidenav/slider-items/slider/slider-item.directive.ts
