+++ src/app/components/sidenav/slider-items/slider/slider.component.scss
