+++ src/app/components/sidenav/slider-items/slider/slider.component.ts
@@ -1,8 +1,8 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { Observable } from 'rxjs';
-import { DriveImageList } from '../../../interfaces/drive';
-import { Slider } from '../../../interfaces/slider';
-import { DriveAPIService } from '../../../services/drive-api.service';
+import { DriveImageList } from '../../../../interfaces/drive';
+import { Slider } from '../../../../interfaces/slider';
+import { DriveAPIService } from '../../../../services/drive-api.service';
 
 @Component({
     selector: 'app-slider',
